surpress warnings in pos_size_tlwh
When children obtain a style before the parent, the parent bin will emit warnings about not being to determine the size/position. This commit adds a check to make sure the parent has a style before it does any calculations for size/position. This as a result will surpress the warnings and will not effect the existing behavior.