use ilmenite ^0.5.1; re-enable default gpu acclerated text
vulkano had a bug where storage buffers where not being detected correctly. As a result a storage buffer in ilmenite was being created as uniform buffer when it should have been a storage buffer. As a result this resulted in issues on nvidia cards. ilmenite now specifies vulkan and spirv version as of 0.5.1 which works around this issue. This should probably have been done anyways. This issue is resolved in https://github.com/vulkano-rs/vulkano/pull/1642