various improvements/fixes
Improved arc-swap usage for BinHrchy. Fixed bug where an update wouldn't trigger when the parent was changed. Fixed bug where text wouldn't appear right away. Fixed bug with ScrollBar where when the content changed and the scroll amount was beyond the content, it wouldn't scroll to an appropriate location. Fixed bug that would cause the OVD to stall when a Bin was dropped or dissociated with a window while an a worker cycle was in progress.