big oops
When vertex data was upload the range was never actually set. Meaning that all vertex data would be uploaded every frame. On top of this bins that were removed or hidden would still persist as it wasn't tracked properly. Upon testing ran into overlapping regions which was sort of expected, but assumed it worked before. It didn't... it just never executed that code. Vertex buffers are now twice as large allowing for each half to be used seperately to avoid overlapping regions. There is still some performance on the table, but a very small amount. It might be possible to avoid swapping halves of the buffer reducing the gpu copies, but gpu's go brrrr.